finalization/requirements.txt
# Requirements for LatentWire training module

# Core dependencies (required)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.60.0
jsonlines>=3.0.0
tokenizers>=0.14.0
pyarrow>=12.0.0

# Evaluation metrics
rouge-score>=0.1.2
nltk>=3.8.0

# Optional but recommended
accelerate>=0.20.0
peft>=0.5.0  # For LoRA support
bitsandbytes>=0.40.0  # For quantization

# Optional: Prompt compression baselines
# llmlingua>=0.2.0  # Uncomment if using LLMLingua baseline

# Development dependencies
pytest>=7.0.0
black>=22.0.0
mypy>=0.950
pyyaml>=6.0

# Optional: Visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
# pandas>=2.0.0
